Portal-Wide Scope and MetaAgent Creation
introduced scope to the registration methods, Portals will no longer
pass the registration on to the Router if the agent only wants to be
Portal-Wide. However Router-Wide currently makes the agent also Global.
Fixed the problem of router's sending a registration message around
making it so every router owns the agent.

Portal and Router now has an insertMetaAgent() method and Routers can
now create Portals and Routers.